Fixed a few warnings.

+++ source/direct3d9/Effect.cpp
@@ -167,7 +167,7 @@ namespace Direct3D9
 		if( data == nullptr )
 		{
 			String^ compilationErrorsLocal;
-			UINT size = static_cast<SIZE_T>( ds->RemainingLength );
+			UINT size = static_cast<UINT>( ds->RemainingLength );
 			Effect^ effect = FromMemory_Internal( device, ds->SeekToEnd(), size, preprocessorDefines, includeFile,
 				skipConstants, flags, pool, &compilationErrorsLocal );
 
@@ -186,7 +186,7 @@ namespace Direct3D9
 		array<Byte>^ data = Utilities::ReadStream( stream, 0, &ds );
 		if( data == nullptr )
 		{
-			UINT size = static_cast<SIZE_T>( ds->RemainingLength );
+			UINT size = static_cast<UINT>( ds->RemainingLength );
 			return FromMemory_Internal( device, ds->SeekToEnd(), size, preprocessorDefines, includeFile,
 				skipConstants, flags, pool, NULL );
 		}
@@ -217,7 +217,7 @@ namespace Direct3D9
 		array<Byte>^ data = Utilities::ReadStream( stream, 0, &ds );
 		if( data == nullptr )
 		{
-			UINT size = static_cast<SIZE_T>( ds->RemainingLength );
+			UINT size = static_cast<UINT>( ds->RemainingLength );
 			return FromMemory_Internal( device, ds->SeekToEnd(), size, nullptr, nullptr,
 				nullptr, flags, nullptr, NULL );
 		}
